ml/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.16.2
flask==2.3.2
prometheus-client==0.17.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
keras==3.13.2